+++ .scripts/deploy-docs.sh
@@ -15,8 +15,10 @@ sed -i "1 s|$| ${TRAVIS_TAG}|" apidoc/introduction.md
 # run apidoc
 apidoc -i . -f js -e node_modules
 
-if [[ -z "${TRAVIS_TAG}" ]]; then
-  echo "Docs can only be built from tags"
+# should contain the current api tag or nothing
+isApiTag=$(echo "${TRAVIS_TAG}" | grep -E '^v[0-9]+$' || true)
+if [[ -z "$isApiTag" ]]; then
+  echo "Docs can only be built from api version tags"
   exit 0
 fi